+++ include/SI/solid_angle.h
@@ -21,19 +21,19 @@ template <typename _type, typename _ratio>
 using solid_angle_t = detail::unit_t<'R', std::ratio<1>, _type, _ratio>;
 
 template <typename _type>
-using atto_sterradian_t = solid_angle_t<_type, std::atto>;
+using atto_steradian_t = solid_angle_t<_type, std::atto>;
 template <typename _type>
-using femto_sterradian_t = solid_angle_t<_type, std::femto>;
+using femto_steradian_t = solid_angle_t<_type, std::femto>;
 template <typename _type>
-using pico_sterradian_t = solid_angle_t<_type, std::pico>;
+using pico_steradian_t = solid_angle_t<_type, std::pico>;
 template <typename _type>
-using nano_sterradian_t = solid_angle_t<_type, std::nano>;
+using nano_steradian_t = solid_angle_t<_type, std::nano>;
 template <typename _type>
-using micro_sterradian_t = solid_angle_t<_type, std::micro>;
+using micro_steradian_t = solid_angle_t<_type, std::micro>;
 template <typename _type>
-using milli_sterradian_t = solid_angle_t<_type, std::milli>;
+using milli_steradian_t = solid_angle_t<_type, std::milli>;
 template <typename _type>
-using sterradian_t = solid_angle_t<_type, std::ratio<1>>;
+using steradian_t = solid_angle_t<_type, std::ratio<1>>;
 
 // specialize unit_symbol for usage with stream operators
 template <>
@@ -51,71 +51,71 @@ struct unit_symbol<'R', _ratio>
 inline namespace literals {
 
 template <char... _digits>
-constexpr atto_sterradian_t<int64_t> operator""_asr() {
-  return atto_sterradian_t<int64_t>{
+constexpr atto_steradian_t<int64_t> operator""_asr() {
+  return atto_steradian_t<int64_t>{
       SI::detail::parsing::Number<_digits...>::value};
 }
 
 template <char... _digits>
-constexpr femto_sterradian_t<int64_t> operator""_fsr() {
-  return femto_sterradian_t<int64_t>{
+constexpr femto_steradian_t<int64_t> operator""_fsr() {
+  return femto_steradian_t<int64_t>{
       SI::detail::parsing::Number<_digits...>::value};
 }
 
 template <char... _digits>
-constexpr pico_sterradian_t<int64_t> operator""_psr() {
-  return pico_sterradian_t<int64_t>{
+constexpr pico_steradian_t<int64_t> operator""_psr() {
+  return pico_steradian_t<int64_t>{
       SI::detail::parsing::Number<_digits...>::value};
 }
 
 template <char... _digits>
-constexpr nano_sterradian_t<int64_t> operator""_nsr() {
-  return nano_sterradian_t<int64_t>{
+constexpr nano_steradian_t<int64_t> operator""_nsr() {
+  return nano_steradian_t<int64_t>{
       SI::detail::parsing::Number<_digits...>::value};
 }
 
 template <char... _digits>
-constexpr micro_sterradian_t<int64_t> operator""_usr() {
-  return micro_sterradian_t<int64_t>{
+constexpr micro_steradian_t<int64_t> operator""_usr() {
+  return micro_steradian_t<int64_t>{
       SI::detail::parsing::Number<_digits...>::value};
 }
 
 template <char... _digits>
-constexpr milli_sterradian_t<int64_t> operator""_msr() {
-  return milli_sterradian_t<int64_t>{
+constexpr milli_steradian_t<int64_t> operator""_msr() {
+  return milli_steradian_t<int64_t>{
       SI::detail::parsing::Number<_digits...>::value};
 }
 
-template <char... _digits> constexpr sterradian_t<int64_t> operator""_sr() {
-  return sterradian_t<int64_t>{SI::detail::parsing::Number<_digits...>::value};
+template <char... _digits> constexpr steradian_t<int64_t> operator""_sr() {
+  return steradian_t<int64_t>{SI::detail::parsing::Number<_digits...>::value};
 }
 
-constexpr atto_sterradian_t<long double> operator""_asr(long double value) {
-  return atto_sterradian_t<long double>{value};
+constexpr atto_steradian_t<long double> operator""_asr(long double value) {
+  return atto_steradian_t<long double>{value};
 }
 
-constexpr femto_sterradian_t<long double> operator""_fsr(long double value) {
-  return femto_sterradian_t<long double>{value};
+constexpr femto_steradian_t<long double> operator""_fsr(long double value) {
+  return femto_steradian_t<long double>{value};
 }
 
-constexpr pico_sterradian_t<long double> operator""_psr(long double value) {
-  return pico_sterradian_t<long double>{value};
+constexpr pico_steradian_t<long double> operator""_psr(long double value) {
+  return pico_steradian_t<long double>{value};
 }
 
-constexpr nano_sterradian_t<long double> operator""_nsr(long double value) {
-  return nano_sterradian_t<long double>{value};
+constexpr nano_steradian_t<long double> operator""_nsr(long double value) {
+  return nano_steradian_t<long double>{value};
 }
 
-constexpr micro_sterradian_t<long double> operator""_usr(long double value) {
-  return micro_sterradian_t<long double>{value};
+constexpr micro_steradian_t<long double> operator""_usr(long double value) {
+  return micro_steradian_t<long double>{value};
 }
 
-constexpr milli_sterradian_t<long double> operator""_msr(long double value) {
-  return milli_sterradian_t<long double>{value};
+constexpr milli_steradian_t<long double> operator""_msr(long double value) {
+  return milli_steradian_t<long double>{value};
 }
 
-constexpr sterradian_t<long double> operator""_sr(long double value) {
-  return sterradian_t<long double>{value};
+constexpr steradian_t<long double> operator""_sr(long double value) {
+  return steradian_t<long double>{value};
 }
 
 } // namespace literals